+++ packages/cloud-toolkit/src/errors.ts
@@ -1,2 +1,8 @@
-// eslint-disable-next-line import/prefer-default-export
+/* eslint-disable max-classes-per-file */
 export class BusinessError extends Error {}
+export class CustomizationError extends BusinessError {
+  constructor(message: string, stack?: string) {
+    super(message);
+    this.stack = stack;
+  }
+}